requirements.txt
Django==1.10.4
django-cors-headers==1.3.1
djangorestframework==3.5.3
psycopg2==2.6.2
